+++ kanidmd/core/Cargo.toml
@@ -11,11 +11,10 @@ license.workspace = true
 homepage.workspace = true
 repository.workspace = true
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 async-std = { workspace = true, features = ["tokio1"] }
 async-trait.workspace = true
+chrono.workspace = true
 compact_jwt.workspace = true
 futures-util.workspace = true
 http-types.workspace = true
@@ -24,7 +23,9 @@ kanidmd_lib.workspace = true
 ldap3_proto.workspace = true
 libc.workspace = true
 openssl.workspace = true
+rand.workspace = true
 regex.workspace = true
+saffron.workspace = true
 serde = { workspace = true, features = ["derive"] }
 serde_json.workspace = true
 sketching.workspace = true
