+++ kanidmd/idm/build.rs
@@ -1,0 +1,5 @@
+// include!("src/lib/audit_loglevel.rs");
+
+fn main() {
+    profiles::apply_profile();
+}
